% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cqe_metrics.R
\name{calculate_cqe_metrics}
\alias{calculate_cqe_metrics}
\title{Calculate CQEs Metrics}
\usage{
calculate_cqe_metrics(
  limma_df,
  species,
  expected_fcs,
  species_protein_annotation
)
}
\arguments{
\item{limma_df}{limma output (dataframe)}

\item{species}{names of species compared (character vector)}

\item{expected_fcs}{expected fold changes for each species (numeric vector)}

\item{species_protein_annotation}{a tibble with 2 columns,
the first column must contain ids matching to the rownames of the limma df (e.g. protein groups)
the second column must identify the species which the ID comes from. This can either be specified directly or supplied at the end of a string, e.g. Uniprot name.
the species must match with those provided in species vector}
}
\value{
A tibble summarising the total number of proteins, "INV_FQR", "Precision" and "BS_FQR" per species.
}
\description{
Calculate metrics to assess controlled quantitative experiments
}
