% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\docType{methods}
\name{qcPlots}
\alias{qcPlots}
\alias{qcPlots,character,data.frame-method}
\title{QC plots for CQE analysis}
\usage{
qcPlots(
  metadata_path,
  expr.df,
  output_dir,
  filename_ext = "",
  columns_per_plot
)

\S4method{qcPlots}{character,data.frame}(
  metadata_path,
  expr.df,
  output_dir,
  filename_ext = "",
  columns_per_plot
)
}
\arguments{
\item{metadata_path}{Path to metadata table. Metadata table should contain a column 'Run' containing
the unique values from the 'Run' column in the initial report, and columns 'Group' and 'Replicate'
from which unique sample labels can be formed. Should be the same metadata file used in
the call to \code{standardProcessing}.}

\item{expr.df}{expression data frame as produced by \code{standardProcessing}. Annotation columns may need to be removed.}

\item{output_dir}{Path to directory where figures will be put.}

\item{filename_ext}{Spliced into into output file names e.g. if filename_ext = "MEDNORM", output files will
be \code{log2protboxplotsMEDNORM_all.jpeg} etc. Defaults to filename_ext = "" to leave the base forms unchanged.}

\item{columns_per_plot}{Used for boxplots}
}
\description{
Produce useful QC plots following pre-processing.
}
\details{
This function produces boxplots, density plots, correlation plots, PCA plots.
}
