% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardProcessing.R
\docType{methods}
\name{standardProcessing}
\alias{standardProcessing}
\alias{standardProcessing,character,ANY,ANY,character-method}
\title{Preprocessing for CQEs}
\usage{
standardProcessing(
  report_path,
  pg.q = 0.01,
  parq_qval_cutoff = 0.01,
  metadata_path,
  lfq_col = "Precursor.Quantity",
  pg_path,
  pr_path,
  min_peptides = 2,
  filter_contaminants = TRUE,
  apply_normalization = TRUE
)

\S4method{standardProcessing}{character,ANY,ANY,character}(
  report_path,
  pg.q = 0.01,
  parq_qval_cutoff = 0.01,
  metadata_path,
  lfq_col = "Precursor.Quantity",
  pg_path,
  pr_path,
  min_peptides = 2,
  filter_contaminants = TRUE,
  apply_normalization = TRUE
)
}
\arguments{
\item{report_path}{Path to initial report (report.parquet)}

\item{pg.q}{diann_matrix protein group q-value threshold}

\item{parq_qval_cutoff}{Q-value threshold for filtering input report passed to diann_maxlfq}

\item{metadata_path}{Path to metadata table. Metadata table should contain a column 'Run' containing
the unique values from the 'Run' column in the initial report, and columns 'Group' and 'Replicate'
from which unique sample labels can be formed}

\item{lfq_col}{Column from diann report file to use in quantitation}

\item{pg_path}{Path to pg file}

\item{pr_path}{Path to pr file}

\item{min_peptides}{Minimum number of peptides per protein (used when filtering protein groups)}

\item{filter_contaminants}{Logical to indicate whether protein groups starting 'Cont_' should be filtered out}

\item{apply_normalization}{Logical to indicate whether limma::normalizeMedianValues should be applied}
}
\value{
A dataframe containing annotation and values per protein group for each sample.
}
\description{
Apply a standard pre-processing pipeline.
}
\details{
This function applies a standard pre-processing pipeline with some
useful customization options#'
}
